Update: Core modules, knowledge, multimodal and robot enhancements
- Updated CMakeLists.txt build configuration
- Enhanced backend service (backend.c) and frontend (main.js)
- Improved cognition: metacognition updates
- Updated concurrency: lock-free data structures
- Enhanced core: CFC network module
- Improved knowledge: self-check and semantic parsing modules
- Updated multimodal: dialogue and OCR processing
- Enhanced reasoning: planning module
- Updated robot control module

+++ src/cognition/metacognition.c
@@ -1438,7 +1438,7 @@ int metacognition_self_correction(MetacognitionSystem* system,
             break;
         case 2: /* 预测误差增加：缩短预测范围、增强反馈 */
             system->prediction_config.prediction_horizon = 
-                (size_t)(system->prediction_config.prediction_horizon * 0.8f);
+                system->prediction_config.prediction_horizon * 0.8f;
             if (system->prediction_config.prediction_horizon < 1)
                 system->prediction_config.prediction_horizon = 1;
             system->model_update_config.learning_rate *= 0.9f;
